# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yudetter SNS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/{username}/yudates:
    get:
      operationId: getUserYudates
      tags: [users]
      summary: Get yudates by a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User yudates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YudatePage"
  /users/{username}/likes:
    get:
      operationId: getUserLikes
      tags: [users]
      summary: Get yudates liked by a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Liked yudates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YudatePage"
...
  /users/{username}/follow/approve:
    post:
      summary: Approve follow request
      operationId: approveFollowRequest
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: User not found
  /users/{username}/follow/reject:
    post:
      summary: Reject follow request
      operationId: rejectFollowRequest
      tags:
        - Users
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "404":
          description: User not found
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
    delete:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollow result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /users/{username}/followers:
    get:
      operationId: getFollowers
      tags: [users]
      summary: Get followers of a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /users/{username}/following:
    get:
      operationId: getFollowing
      tags: [users]
      summary: Get users a user is following
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Following list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"

  # ── EXPLORE ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ActionResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    UserProfile:
      type: object
      properties:
        setupComplete:
          type: boolean
        isPrivate:
          type: boolean
        isBlocking:
          type: boolean
        isBlockedBy:
          type: boolean
        pinnedYudateId:
          type: ["integer", "null"]
        headerUrl:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        followerCount:
          type: integer
        followingCount:
          type: integer
        yudateCount:
          type: integer
        isFollowing:
          type: boolean
        isFollowPending:
          type: boolean
        createdAt:
          type: string
          format: date-time
        visibility:
          type: string
          enum: [public, followers]
        isSpoiler:
          type: boolean
        yudedollar:
          type: integer
        badgeType:
          type: ["string", "null"]
        consecutiveLoginDays:
          type: integer
        rankingOptIn:
          type: boolean
      required:
        - id
        - clerkId
        - username
        - displayName
        - email
        - followerCount
        - followingCount
        - yudateCount
        - isFollowing
        - isFollowPending
        - isPrivate
        - isBlocking
        - isBlockedBy
        - setupComplete
        - createdAt
        - yudedollar
        - consecutiveLoginDays
        - rankingOptIn
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        headerUrl:
          type: ["string", "null"]
        isPrivate:
          type: boolean
...
    UserPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        nextCursor:
          type: ["integer", "null"]
      required:
        - items

    # ── YUDATE ──────────────────────────────────────────────────────────────
    Yudate:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        likeCount:
          type: integer
        reyudateCount:
          type: integer
        replyCount:
          type: integer
        isLiked:
          type: boolean
        isReyudated:
          type: boolean
        quotedYudate:
          oneOf:
            - $ref: "#/components/schemas/QuotedYudate"
            - type: "null"
        replyToId:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        visibility:
          type: string
          enum: [public, followers]
        isSpoiler:
          type: boolean
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/YudateReaction"
        superYudateAmount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - likeCount
        - reyudateCount
        - replyCount
        - isLiked
        - isReyudated
        - superYudateAmount
        - createdAt
    QuotedYudate:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          $ref: "#/components/schemas/UserProfile"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - author
        - createdAt
...
    YudatePage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Yudate"
        nextCursor:
          type: ["integer", "null"]
      required:
        - items

    # ── NOTIFICATION ─────────────────────────────────────────────────────────
...
    YudateReaction:
      type: object
      properties:
        emoji:
          type: string
        count:
          type: integer
        isReacted:
          type: boolean
      required:
        - emoji
        - count
        - isReacted
...